requirements.txt
alembic==1.13.3
fastapi==0.115.4
httpx==0.27.2
jinja2==3.1.4
pydantic-settings==2.6.1
python-multipart==0.0.17
uvicorn==0.32.0